+++ projectroles/forms.py
@@ -167,7 +167,7 @@ class ProjectForm(forms.ModelForm):
                             **setting_kwargs
                         )
 
-                        # Set initial value
+                    # Set initial value
                     if self.instance.pk:
                         self.initial[
                             s_field
@@ -234,6 +234,7 @@ class ProjectForm(forms.ModelForm):
             # Set hidden project field for autocomplete
             self.initial['project'] = self.instance
 
+            # Set owner value but hide the field (updating via member form)
             self.initial['owner'] = self.instance.get_owner().user.sodar_uuid
             self.fields['owner'].widget = forms.HiddenInput()
 
@@ -318,25 +319,31 @@ class ProjectForm(forms.ModelForm):
                 plugin, APP_SETTING_SCOPE_PROJECT, user_modifiable=True
             )
 
-            for s_key in p_settings:
-                s = p_settings[s_key]
+            for s_key, s_val in p_settings.items():
                 s_field = 'settings.{}.{}'.format(plugin.name, s_key)
 
-                if s['type'] == 'JSON':
-                    # for some reason, there is a distinct possiblity, that the initial value has been discarded and we get '' as value. Seems to only happen in automated tests. Will catch that here.
-                    if self.cleaned_data[s_field] == '':
+                if s_val['type'] == 'JSON':
+                    # for some reason, there is a distinct possiblity, that the
+                    # initial value has been discarded and we get '' as value.
+                    # Seems to only happen in automated tests. Will catch that
+                    # here.
+                    if not self.cleaned_data.get(s_field):
                         self.cleaned_data[s_field] = '{}'
+
                     try:
+
                         self.cleaned_data[s_field] = json.loads(
-                            self.cleaned_data[s_field]
+                            self.cleaned_data.get(s_field)
                         )
+
                     except json.JSONDecodeError as err:
+                        # TODO: Shouldn't we use add_error() instead?
                         raise forms.ValidationError(
-                            'Couldn\'t encode json.\n' + str(err)
+                            'Couldn\'t encode JSON\n' + str(err)
                         )
 
                 if not self.app_settings.validate_setting(
-                    setting_type=s['type'],
+                    setting_type=s_val['type'],
                     setting_value=self.cleaned_data.get(s_field),
                 ):
                     self.add_error(s_field, 'Invalid value')